requirements.txt
Bootstrap-Flask==2.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
defusedxml==0.7.1
Flask==2.2.3
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
m3u8==3.4.0
MarkupSafe==2.1.2
psutil==5.9.4
python-vlc==3.0.18121
PyYAML==6.0
requests==2.28.2
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
